转速转换错误问题解决，代码逻辑错误修改，细节问题处理完毕，目前就差解决OLED连接电机驱动模块黑屏花屏问题初级系统即可完成

+++ demo9-x 智能温控风扇系统/demo9-x1 串口发送温控信号+PWM占空比/Hardware/Motor.h
@@ -4,7 +4,7 @@
 
 void Motor_Init(void);
 
-void Motor_SetSpeed(int8_t Speed);
+void Motor_SetSpeed(int16_t Speed);
 
 void Motor_Turn(void);
