+++ demo9-x 智能温控风扇系统/demo9-x1 串口发送温控信号+PWM占空比/User/FanModule.h
@@ -1,9 +1,9 @@
 #ifndef __TEMPMODULE__H
 #define __TEMPMODULE__H
 
-uint16_t FanModule(uint8_t KeyNum);
-void AutoMode(uint8_t KeyNum,uint16_t Temp);
+int16_t FanModule(uint8_t KeyNum);
+void AutoMode(uint8_t KeyNum,float Temp);
 
-extern uint16_t AimTemp;
+extern float AimTemp;
 
 #endif
